requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
Django==3.2.7
django-filter==2.4.0
django-smtp-ssl==1.0
djangorestframework==3.12.4
gunicorn==20.1.0
kombu==5.1.0
Markdown==3.3.4
prompt-toolkit==3.0.20
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==3.0.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
